chore: change a zondax/hid version for mac os
### Description
The build issue below occurs on macOS 12.0 or higher when installing libraries with "make install" command.
It's occurred because the "kIOMasterPortDefault" is deprecated in macOS 12.0.

- warning: 'kIOMasterPortDefault' is deprecated: first deprecated in macOS 12.0 [-Wdeprecated-declarations]

### Related Issues
Here are links to similar issues in other repositories.

- [https://github.com/kovidgoyal/kitty/issues/3719](https://github.com/kovidgoyal/kitty/issues/3719)
- [https://github.com/datreeio/datree/issues/512](https://github.com/datreeio/datree/issues/512)

### Solution
The issue is related to "zondax/hid v0.9.0" among the installed libraries. On the [zondax/hid releases page](https://github.com/Zondax/hid/releases), the version v0.9.1 is released for macOS 12.0 or higher.
So, I changed the version of zondax/hid in the "go.mod" file to v0.9.1.